Big changes!
— Added the ability to stack draw twos/draw fours and either force
players to draw the running total or to play a draw two/draw four.
— Added a prompt for users who either need to draw x amount of cards or
play a draw two/draw four.
— Added additional functionality where if a player is required to draw
x amount of cards, clicking on the “Draw Card” button will
automatically draw x amount of cards for them (same with the “Draw 2”
and “Draw 4” buttons. If a player clicks either of those buttons and
they don’t *need* to draw that many, they will be given that number
regardless).
— Disallowed players from drawing when not their turn.
— Disallowed players from starting a new game in the middle of an
ongoing game.
— Added function comments.